# car-models-service-core/docs/api-docs.yml
openapi: 3.0.1
info:
  title: Car models service API
  description: The API to persist and search car models in a database
  version: "1.0"
servers:
- url: http://localhost:8180
  description: Generated server url
paths:
...
  /v1/manufacturers:
    get:
      tags:
      - manufacturer
      summary: Get all manufacturers
      description: Retrieve all manufacturers from the database
      operationId: getAllManufacturers
      parameters:
      - name: page
        in: query
        description: Zero-based page index (0..N)
        required: false
        schema:
          minimum: 0
          type: integer
          default: 0
      - name: size
        in: query
        description: The size of the page to be returned
        required: false
        schema:
          minimum: 1
          type: integer
          default: 100
      - name: sort
        in: query
        description: "Sorting criteria in the format: property,(asc|desc). Default\
          \ sort order is ascending. Multiple sort criteria are supported."
        required: false
        schema:
          type: array
          items:
            type: string
      responses:
        "405":
          description: Method Not Allowed
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "200":
          description: Ok
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PageManufacturerDto'
      security:
      - bearerAuth: []
    post:
      tags:
      - manufacturer
      summary: Save a manufacturer
      description: Search for such manufacturer in the database if it is missing then
        persist it
      operationId: saveManufacturer
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ManufacturerDto'
            examples:
              manufacturer:
                description: manufacturer
                value:
                  name: Audi
        required: true
      responses:
        "405":
          description: Method Not Allowed
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "201":
          description: Created
          headers:
            Location:
              description: "/v1/manufacturers/{name}"
              style: simple
              schema:
                type: string
        "409":
          description: Conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
      - bearerAuth: []
...
  /v1/manufacturers/{name}:
    get:
      tags:
      - manufacturer
      summary: Get a manufacturer by its name
      description: Search for and retrieve a manufacturer from the database
      operationId: getManufacturerByName
      parameters:
      - name: name
        in: path
        required: true
        schema:
          type: string
      responses:
        "405":
          description: Method Not Allowed
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "200":
          description: Ok
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ManufacturerDto'
              examples:
                manufacturer:
                  description: manufacturer
                  value:
                    name: Audi
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
      - bearerAuth: []
    delete:
      tags:
      - manufacturer
      summary: Delete a manufacturer by its name
      description: Seach for and delete a manufacturer from the database by its name
      operationId: deleteManufacturerByName
      parameters:
      - name: name
        in: path
        required: true
        schema:
          type: string
      responses:
        "405":
          description: Method Not Allowed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "204":
          description: No Content
      security:
      - bearerAuth: []
...
components:
  schemas:
    ErrorResponse:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
        status:
          type: integer
          format: int32
        error:
          type: string
        message:
          type: string
        path:
          type: string
        violations:
          type: array
          items:
            $ref: '#/components/schemas/Violation'
    Violation:
      type: object
      properties:
        attribute:
          type: string
        message:
          type: string
...
    ManufacturerDto:
      required:
      - name
      type: object
      properties:
        name:
          type: string
...
    PageableObject:
      type: object
      properties:
        pageNumber:
          type: integer
          format: int32
        pageSize:
          type: integer
          format: int32
        offset:
          type: integer
          format: int64
        sort:
          $ref: '#/components/schemas/SortObject'
        paged:
          type: boolean
        unpaged:
          type: boolean
    SortObject:
      type: object
      properties:
        sorted:
          type: boolean
        empty:
          type: boolean
        unsorted:
          type: boolean
...
    PageManufacturerDto:
      type: object
      properties:
        totalPages:
          type: integer
          format: int32
        totalElements:
          type: integer
          format: int64
        pageable:
          $ref: '#/components/schemas/PageableObject'
        first:
          type: boolean
        last:
          type: boolean
        number:
          type: integer
          format: int32
        size:
          type: integer
          format: int32
        content:
          type: array
          items:
            $ref: '#/components/schemas/ManufacturerDto'
        sort:
          $ref: '#/components/schemas/SortObject'
        numberOfElements:
          type: integer
          format: int32
        empty:
          type: boolean
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT